Moved DiscordWebhook to discord package
Logging error messages from hook
Fixed user-agent
Replaced json.simple with gson
